prov/tcp: Convert ssize_t return type to int
The tcp provider was initially using ssize_t return types because
the ofi_bsock_* functions were returning ssize_t values as well.

Since the bsock interface now exclusively uses int for the return type,
we can convert all the ssize_t return types from the provider to int.

+++ prov/tcp/src/tcpx.h
@@ -223,7 +223,7 @@ struct tcpx_cur_rx {
 	size_t			hdr_done;
 	size_t			data_left;
 	struct tcpx_xfer_entry	*entry;
-	ssize_t			(*handler)(struct tcpx_ep *ep);
+	int			(*handler)(struct tcpx_ep *ep);
 };
 
 struct tcpx_cur_tx {
